wims api bug fixes

+++ src/main/java/com/ascentdev/wims/entity/RackEntity.java
@@ -61,5 +61,5 @@ public class RackEntity {
   long txnMawbId;
   
   @Column(name="txn_hawb_id")
-  int txnHawbId;
+  long txnHawbId;
 }
